Split project into backend and frontend folders
(where frontend has a subfolder cli, as we're planning on different frontends already
