+++ BUILD.bak
@@ -1,9 +1,16 @@
-java_binary(
+java_library(
     name = "LineCounter",
     srcs = glob(["src/main/java/net/darkseraphim/linecount/**"]),
-    main_class = "net.darkseraphim.linecount.LineCounter",
     deps = [
         "@maven//:org_slf4j_slf4j_api",
         "@maven//:org_slf4j_slf4j_simple",
     ],
+)
+
+java_binary(
+    name = "LineCounterCLI",
+    main_class = "src/main/java/net/darkseraphim/linecount/LineCounter.java",
+    deps = [
+        ":LineCounter" 
+    ]
 )
