+++ backend/BUILD
@@ -1,0 +1,8 @@
+java_library(
+    name = "LineCounter",
+    srcs = glob(["src/main/java/net/darkseraphim/linecount/**"]),
+    deps = [
+        "@maven//:org_slf4j_slf4j_api",
+    ],
+    visibility = ["//visibility:public"]
+)
