+++ backend/src/main/java/net/darkseraphim/linecount/LineEnding.java
