+++ backend/src/main/java/net/darkseraphim/linecount/Utils.java
