+++ backend/src/main/java/net/darkseraphim/linecount/ex/LineCountException.java
