+++ backend/src/main/java/net/darkseraphim/linecount/strategy/AsyncLineCountStrategy.java
