+++ backend/src/main/java/net/darkseraphim/linecount/strategy/LineCountStrategy.java
