+++ backend/src/main/java/net/darkseraphim/linecount/strategy/LineEndingDetectionStrategy.java
