+++ backend/src/main/java/net/darkseraphim/linecount/strategy/content/ParallelDelegatingLineCountStrategy.java
