+++ backend/src/main/java/net/darkseraphim/linecount/strategy/delegating/AsyncDelegatingLineCountStrategy.java
