+++ backend/src/main/java/net/darkseraphim/linecount/strategy/lineendingdetection/OsBasedLineEndingDetection.java
