+++ backend/src/main/java/net/darkseraphim/linecount/supplier/AbstractLinesSupplier.java
