+++ backend/src/main/java/net/darkseraphim/linecount/supplier/IterableBasedLinesSupplier.java
